requirements.txt
Flask == 3.0.2
gunicorn == 21.2.0
qrcode == 7.4.2
validators == 0.23.1
google-cloud-storage == 2.16.0
pillow == 10.2.0
